Remove schemas and old pblite
