+++ hangups/parsers.py
@@ -4,7 +4,7 @@ import logging
 from collections import namedtuple
 import datetime
 
-from hangups import javascript, user, hangouts_pb2, pblite2
+from hangups import javascript, user, hangouts_pb2, pblite
 
 
 logger = logging.getLogger(__name__)
@@ -49,7 +49,7 @@ def _parse_payload(payload):
         # This is a BatchUpdate containing StateUpdate messages
         batch_update = hangouts_pb2.BatchUpdate()
         # TODO: error handling
-        pblite2.decode(batch_update, payload)
+        pblite.decode(batch_update, payload)
         for state_update in batch_update.state_update:
             logger.debug('Received StateUpdate:\n%s', state_update)
             yield state_update
